LayoutTests/platform/ios-simulator/css3/device-adapt/opera/syntax-001-expected.txt


FAIL CSS Test: @viewport syntax - inside @media assert_equals: expected 500 but got 800(stack: assert@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:2168:73
assert_equals@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:705:15
file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/css3/device-adapt/opera/syntax-001.xhtml:69:22
step@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:1186:30
onload@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/css3/device-adapt/opera/syntax-001.xhtml:68:14)
